\name{test_minimal}
\alias{test_minimal}
\title{Test Minimal Function}
\description{
A minimal test function.
}
\usage{
test_minimal(x, y = 1)
}
\arguments{
\item{x}{First argument.}
\item{y}{Second argument with default.}
}
\details{
Some details here.

More details.
}
\value{
Returns a value.
}
\examples{
\dontrun{
test_minimal(5, 10)
}
}
